+++ CHANGELOG.md
@@ -37,6 +37,10 @@ Section Order:
 
 <!-- Your changes go here -->
 
+### Changed
+
+- Modular providers
+
 ## [3.2.5] - 2026-05-05
 
 ### Changed
