+++ fleetfinder/templatetags/fleetfinder.py
@@ -9,10 +9,9 @@ from django.template.defaulttags import register
 from allianceauth.services.hooks import get_extension_logger
 
 # AA Fleet Finder
-from fleetfinder import __title__
-from fleetfinder.providers import AppLogger
+from fleetfinder.providers.applogger import AppLogger
 
-logger = AppLogger(my_logger=get_extension_logger(__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(__name__))
 
 
 @register.filter
